Fix for issue #314 in Firefox and other browsers
In firefox, and other earlier browsers it seems as though they haven't fully implemented the map functionality. Through setting the index of the Col, this should allow it to access the correct variables inside the object.
Whilst also experimenting, I found that setting the Date from a string wasn't working and was only supported in Chrome. As you can see, I've manually set it up to put the year, month and day in, rather than doing it from the string.